Crate libra_secure_push_metrics

Macros

define_counters
Defines a list of counters with help messages Additionally, requires a prefix which will be prepended to all counters

Structs

Counter
A Counter is a cumulative metric that represents a single monotonically increasing counter whose value can only increase or be reset to zero on restart.

Gauge
A Gauge is a metric that represents a single numerical value that can arbitrarily go up and down.

MetricsPusher
MetricsPusher provides a function to push a list of Metrics to a configurable pushgateway endpoint.

Traits

Metrics

PushMetricFormat
PushMetricFormat contains a single function which converts the metric into a string format which can be pushed to Prometheus