Macro libra_secure_push_metrics::define_counters

macro_rules! define_counters {
    [
        $prefix:expr,$(($counter:ident: $countertype:ty, $help: expr)),* $(,)?
    ] => { ... };
}

Defines a list of counters with help messages Additionally, requires a prefix which will be prepended to all counters

Example of defining two counters and using the prefix libra_safety_rules for counter names

Also defines a static COUNTERS which contains a reference to all the defined counters

define_counters![
    "libra_safety_rules",
    (
        sign_proposal: Counter,
        "sign_proposal counter counts sign_proposals"
    ),
    (
        sign_timeout: Counter,
        "sign_timeout counter counts sign_timeouts"
    ),
    (some_gauge: Gauge, "example help for a gauge metric"),
];

lazy_static! {
    pub static ref COUNTERS: Arc<Counters> = Arc::new(Counters::new());
}

Example of setting them in code

COUNTERS.some_gauge.set(42)
COUNTERS.some_counter.inc()